requirements.txt
python-dateutil==2.9.0
django==4.2.13